Refactor Code Playground component
Refactored the `CodePlayground` component into smaller, more manageable files.  Removed unused imports and files.
