+++ src/hooks/useCodeExecution.ts
@@ -1,0 +1,48 @@
+
+import { useState, useRef } from 'react';
+import { executeJavaScript, executeHTML } from '@/utils/codeExecutor';
+import { useToast } from '@/hooks/use-toast';
+
+export const useCodeExecution = () => {
+  const [output, setOutput] = useState<string>('');
+  const iframeRef = useRef<HTMLDivElement>(null);
+  const { toast } = useToast();
+
+  const handleRun = async (code: string, language: string) => {
+    try {
+      if (language === 'html') {
+        if (iframeRef.current) {
+          const iframe = executeHTML(code);
+          iframeRef.current.innerHTML = '';
+          iframeRef.current.appendChild(iframe);
+        }
+        return;
+      }
+
+      const { result, error, logs = [] } = await executeJavaScript(code);
+      const outputText = [
+        ...(logs.length > 0 ? logs : []),
+        ...(result !== undefined ? [result] : []),
+        ...(error ? [`Error: ${error}`] : [])
+      ].join('\n');
+      
+      setOutput(outputText);
+      
+      toast({
+        description: error ? "Execution failed" : "Code executed successfully",
+        variant: error ? "destructive" : "default",
+      });
+    } catch (err) {
+      toast({
+        description: "Failed to execute code",
+        variant: "destructive",
+      });
+    }
+  };
+
+  return {
+    output,
+    iframeRef,
+    handleRun
+  };
+};
